# emas/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.ProcessSteps:
    properties:
      allow_parallel_execution:
        type: boolean
      batch_size:
        description: 0 = no batch constraint
        type: integer
      default_changeover_time:
        description: minutes
        type: integer
      default_cleaning_time:
        description: minutes
        type: integer
      default_preparation_time:
        description: minutes
        type: integer
      default_processing_time:
        description: minutes
        type: integer
      is_batch_process:
        description: true if step runs in batches
        type: boolean
      machine_type_required:
        type: string
      max_parallel_machines:
        type: integer
      min_batch_size:
        description: minimum batch when splitting
        type: integer
      min_split_qty:
        type: integer
      min_wait_minutes:
        description: e.g. cooling time before next step
        type: integer
      notes:
        type: string
      predecessor_step_ids:
        description: JSON array of step_ids; empty = infer from StepSequence
        type: string
      process_id:
        type: string
      quality_check_required:
        type: boolean
      step_id:
        type: string
      step_name:
        type: string
      step_sequence:
        type: integer
      step_type:
        description: matches reference_step_types.name
        type: string
      transfer_batch_size:
        type: integer
      transfer_minutes:
        description: transport time to next step
        type: integer
    type: object
...
  domain.ProductProcess:
    properties:
      description:
        type: string
      effectiveFrom:
        type: string
      effectiveTo:
        type: string
      isPrimary:
        description: primary vs alternative routing
        type: boolean
      processID:
        type: string
      processName:
        type: string
      productID:
        type: string
      sequence:
        description: order when multiple (0=primary)
        type: integer
      version:
        type: integer
    type: object
...
  emas_internal_handler_dto.CreateProcessRequest:
    properties:
      description:
        type: string
      process_id:
        type: string
      process_name:
        type: string
      product_id:
        type: string
      version:
        type: integer
    required:
    - process_id
    - process_name
    - product_id
    type: object
  emas_internal_handler_dto.CreateProcessStepRequest:
    properties:
      allow_parallel_execution:
        type: boolean
      default_changeover_time:
        type: integer
      default_cleaning_time:
        type: integer
      default_preparation_time:
        type: integer
      default_processing_time:
        type: integer
      machine_type_required:
        type: string
      max_parallel_machines:
        type: integer
      min_split_qty:
        type: integer
      notes:
        type: string
      quality_check_required:
        type: boolean
      step_id:
        type: string
      step_name:
        type: string
      step_sequence:
        type: integer
      step_type:
        type: string
      transfer_batch_size:
        type: integer
    required:
    - step_name
    type: object
...
  emas_internal_handler_dto.Response:
    properties:
      data: {}
      error:
        type: string
      success:
        type: boolean
    type: object
...
  service.AIRecommendation:
    properties:
      action:
        type: string
      icon:
        type: string
      severity:
        type: string
      title:
        type: string
    type: object
  service.ConfidenceSummary:
    properties:
      confidence_pct:
        type: number
      last_trained:
        type: string
      model:
        type: string
    type: object
  service.HighRiskJobPrediction:
    properties:
      delay_minutes:
        type: integer
      issue:
        type: string
      job_id:
        type: string
      machine_name:
        type: string
      risk_level:
        type: string
      risk_score:
        type: number
    type: object
host: localhost:8080
info:
  contact: {}
  description: Factory operations API.
  title: eMas Factory API
  version: "1.0"
paths:
...
  /predictive/confidence:
    get:
      consumes:
      - application/json
      description: Confidence
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/emas_internal_handler_dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.ConfidenceSummary'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/emas_internal_handler_dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/emas_internal_handler_dto.Response'
      summary: Confidence
      tags:
      - predictive
  /predictive/forecast:
    get:
      consumes:
      - application/json
      description: Forecast
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/emas_internal_handler_dto.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/emas_internal_handler_dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/emas_internal_handler_dto.Response'
      summary: Forecast
      tags:
      - predictive
  /predictive/high-risk-jobs:
    get:
      consumes:
      - application/json
      description: List high-risk jobs
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/emas_internal_handler_dto.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/service.HighRiskJobPrediction'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/emas_internal_handler_dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/emas_internal_handler_dto.Response'
      summary: List high-risk jobs
      tags:
      - predictive
  /predictive/recommendations:
    get:
      consumes:
      - application/json
      description: List recommendations
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/emas_internal_handler_dto.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/service.AIRecommendation'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/emas_internal_handler_dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/emas_internal_handler_dto.Response'
      summary: List recommendations
      tags:
      - predictive
...
  /processes:
    get:
      consumes:
      - application/json
      description: List processes
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/emas_internal_handler_dto.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.ProductProcess'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/emas_internal_handler_dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/emas_internal_handler_dto.Response'
      summary: List processes
      tags:
      - process
    post:
      consumes:
      - application/json
      description: Create a process
      parameters:
      - description: Create Process Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/emas_internal_handler_dto.CreateProcessRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/emas_internal_handler_dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.ProductProcess'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/emas_internal_handler_dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/emas_internal_handler_dto.Response'
      summary: Create a process
      tags:
      - process
  /processes/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a process
      parameters:
      - description: Process ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/emas_internal_handler_dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/emas_internal_handler_dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/emas_internal_handler_dto.Response'
      summary: Delete a process
      tags:
      - process
    get:
      consumes:
      - application/json
      description: Get a process by ID
      parameters:
      - description: Process ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/emas_internal_handler_dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.ProductProcess'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/emas_internal_handler_dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/emas_internal_handler_dto.Response'
      summary: Get a process by ID
      tags:
      - process
  /processes/{id}/steps:
    get:
      consumes:
      - application/json
      description: List steps by process ID
      parameters:
      - description: Process ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/emas_internal_handler_dto.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.ProcessSteps'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/emas_internal_handler_dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/emas_internal_handler_dto.Response'
      summary: List steps by process ID
      tags:
      - process
    post:
      consumes:
      - application/json
      description: Add a step to a process
      parameters:
      - description: Process ID
        in: path
        name: id
        required: true
        type: string
      - description: Create Process Step Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/emas_internal_handler_dto.CreateProcessStepRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/emas_internal_handler_dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.ProcessSteps'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/emas_internal_handler_dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/emas_internal_handler_dto.Response'
      summary: Add a step to a process
      tags:
      - process
  /processes/product/{id}:
    get:
      consumes:
      - application/json
      description: Get a process by product ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/emas_internal_handler_dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.ProductProcess'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/emas_internal_handler_dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/emas_internal_handler_dto.Response'
      summary: Get a process by product ID
      tags:
      - process
...
swagger: "2.0"